+++ tencentcloud/cat/v20180409/errors.go
@@ -47,6 +47,12 @@ const (
 	// 资源不存在。
 	FAILEDOPERATION_RESOURCENOTFOUND = "FailedOperation.ResourceNotFound"
 
+	// 发送授权请求失败。
+	FAILEDOPERATION_SENDREQUEST = "FailedOperation.SendRequest"
+
+	// 标签验证错误。
+	FAILEDOPERATION_TAGREQUIREDVERIFYFAILED = "FailedOperation.TagRequiredVerifyFailed"
+
 	// 任务未运行。
 	FAILEDOPERATION_TASKNOTRUNNING = "FailedOperation.TaskNotRunning"
 
@@ -61,6 +67,9 @@ const (
 
 	// 试用任务量超时。
 	FAILEDOPERATION_TRIALTASKEXCEED = "FailedOperation.TrialTaskExceed"
+
+	// json解析失败。
+	FAILEDOPERATION_UNMARSHALRESPONSE = "FailedOperation.UnmarshalResponse"
 
 	// 内部错误。
 	INTERNALERROR = "InternalError"