+++ tencentcloud/ess/v20201111/client.go
@@ -652,7 +652,7 @@ func NewCreateFlowApproversResponse() (response *CreateFlowApproversResponse) {
 // CreateFlowApprovers
 // 补充签署流程本企业签署人信息
 //
-// 适用场景：在通过模版或者文件发起合同时，若未指定本企业签署人信息，则流程发起后，可以调用此接口补充签署人。
+// 适用场景：在通过模板或者文件发起合同时，若未指定本企业签署人信息，则流程发起后，可以调用此接口补充签署人。
 //
 // 同一签署人可以补充多个员工作为候选签署人,最终签署人取决于谁先领取合同完成签署。
 //
@@ -736,7 +736,7 @@ func (c *Client) CreateFlowApprovers(request *CreateFlowApproversRequest) (respo
 // CreateFlowApprovers
 // 补充签署流程本企业签署人信息
 //
-// 适用场景：在通过模版或者文件发起合同时，若未指定本企业签署人信息，则流程发起后，可以调用此接口补充签署人。
+// 适用场景：在通过模板或者文件发起合同时，若未指定本企业签署人信息，则流程发起后，可以调用此接口补充签署人。
 //
 // 同一签署人可以补充多个员工作为候选签署人,最终签署人取决于谁先领取合同完成签署。
 //
@@ -1216,7 +1216,7 @@ func NewCreateFlowSignReviewResponse() (response *CreateFlowSignReviewResponse) 
 //
 // 适用场景: 
 //
-// 在通过接口(CreateFlow 或者CreateFlowByFiles)创建签署流程时，若指定了参数 NeedSignReview 为true,则可以调用此接口提交企业内部签署审批结果。
+// 在通过接口(CreateFlow 或者CreateFlowByFiles)创建签署流程时，若指定了参数 NeedSignReview 为true，且发起方企业作为签署方参与了流程签署，则可以调用此接口提交企业内部签署审批结果。
 //
 // 若签署流程状态正常，且本企业存在签署方未签署，同一签署流程可以多次提交签署审批结果，签署时的最后一个“审批结果”有效。
 //
@@ -1251,7 +1251,7 @@ func (c *Client) CreateFlowSignReview(request *CreateFlowSignReviewRequest) (res
 //
 // 适用场景: 
 //
-// 在通过接口(CreateFlow 或者CreateFlowByFiles)创建签署流程时，若指定了参数 NeedSignReview 为true,则可以调用此接口提交企业内部签署审批结果。
+// 在通过接口(CreateFlow 或者CreateFlowByFiles)创建签署流程时，若指定了参数 NeedSignReview 为true，且发起方企业作为签署方参与了流程签署，则可以调用此接口提交企业内部签署审批结果。
 //
 // 若签署流程状态正常，且本企业存在签署方未签署，同一签署流程可以多次提交签署审批结果，签署时的最后一个“审批结果”有效。
 //