+++ tencentcloud/ess/v20201111/models.go
@@ -961,8 +961,8 @@ type CreateFlowRequestParams struct {
 	// 合同显示的页卡模板，说明：只支持{合同名称}, {发起方企业}, {发起方姓名}, {签署方N企业}, {签署方N姓名}，且N不能超过签署人的数量，N从1开始
 	CustomShowMap *string `json:"CustomShowMap,omitempty" name:"CustomShowMap"`
 
-	// 发起方企业的签署人进行签署操作是否需要企业内部审批。
-	// 若设置为true,审核结果需通过接口 CreateFlowSignReview 通知电子签，审核通过后，发起方企业签署人方可进行签署操作，否则会阻塞其签署操作。
+	// 发起方企业的签署人进行签署操作是否需要企业内部审批。使用此功能需要发起方企业有参与签署。
+	// 若设置为true，审核结果需通过接口 CreateFlowSignReview 通知电子签，审核通过后，发起方企业签署人方可进行签署操作，否则会阻塞其签署操作。
 	// 
 	// 注：企业可以通过此功能与企业内部的审批流程进行关联，支持手动、静默签署合同。
 	NeedSignReview *bool `json:"NeedSignReview,omitempty" name:"NeedSignReview"`
@@ -1014,8 +1014,8 @@ type CreateFlowRequest struct {
 	// 合同显示的页卡模板，说明：只支持{合同名称}, {发起方企业}, {发起方姓名}, {签署方N企业}, {签署方N姓名}，且N不能超过签署人的数量，N从1开始
 	CustomShowMap *string `json:"CustomShowMap,omitempty" name:"CustomShowMap"`
 
-	// 发起方企业的签署人进行签署操作是否需要企业内部审批。
-	// 若设置为true,审核结果需通过接口 CreateFlowSignReview 通知电子签，审核通过后，发起方企业签署人方可进行签署操作，否则会阻塞其签署操作。
+	// 发起方企业的签署人进行签署操作是否需要企业内部审批。使用此功能需要发起方企业有参与签署。
+	// 若设置为true，审核结果需通过接口 CreateFlowSignReview 通知电子签，审核通过后，发起方企业签署人方可进行签署操作，否则会阻塞其签署操作。
 	// 
 	// 注：企业可以通过此功能与企业内部的审批流程进行关联，支持手动、静默签署合同。
 	NeedSignReview *bool `json:"NeedSignReview,omitempty" name:"NeedSignReview"`